refactor aqua helpers into their own package
