+++ controllers/aquascanneraccount_controller.go
@@ -17,14 +17,7 @@ limitations under the License.
 package controllers
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"fmt"
-	"html/template"
-	"io/ioutil"
-	"net/http"
-	"os"
 	"strings"
 	"time"
 
@@ -32,13 +25,13 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	mamoadevopsgovbccav1alpha1 "github.com/bcgov-platform-services/aqua-scan-cli-operator/api/v1alpha1"
+	aqua "github.com/bcgov-platform-services/aqua-scan-cli-operator/aqua_helpers"
 	"github.com/bcgov-platform-services/aqua-scan-cli-operator/utils"
 	"github.com/m1/go-generate-password/generator"
 )
@@ -49,29 +42,6 @@ const aquaScannerAccountFinalizer = "mamoa.devops.gov.bc.ca.devops.gov.bc.ca/fin
 type AquaScannerAccountReconciler struct {
 	client.Client
 	Scheme *runtime.Scheme
-}
-
-type ApplicationScope struct {
-	Name               string
-	NamespacePrefix    string
-	Description        string
-	TechnicalLeadEmail string
-}
-
-type Role struct {
-	Name        string
-	Description string
-	ApplicationScope
-}
-
-type User struct {
-	Name string
-	Role
-	Password string
-}
-
-type aquaResponseJson struct {
-	Message string `json:"message"`
 }
 
 //+kubebuilder:rbac:groups=mamoa.devops.gov.bc.ca.devops.gov.bc.ca,resources=aquascanneraccounts,verbs=get;list;watch;create;update;patch;delete
@@ -162,20 +132,6 @@ func (r *AquaScannerAccountReconciler) Reconcile(ctx context.Context, req ctrl.R
 			return ctrl.Result{Requeue: true}, err
 		}
 
-		// if there are existing accounts with the same name in aqua they must be deleted first
-		accountExists, aquaResErr := doesAquaAccountAlreadyExist(ctrl.Log, aquaScannerAccountName)
-
-		if aquaResErr != nil {
-			return ctrl.Result{Requeue: true}, aquaResErr
-		}
-
-		if accountExists {
-			// only need to delete account becuase we are regenerating a password
-			delErr := deleteAquaAccount(ctrl.Log, aquaScannerAccountName)
-			if delErr != nil {
-				return ctrl.Result{Requeue: true}, delErr
-			}
-		}
 		namespacePrefix := strings.TrimSuffix(req.Namespace, "-test")
 
 		var namespace corev1.Namespace
@@ -190,63 +146,103 @@ func (r *AquaScannerAccountReconciler) Reconcile(ctx context.Context, req ctrl.R
 
 		technicalLeadEmail := utils.GetTechnicalContactFromAnnotation(contactAnnotation)
 
-		applicationScope := ApplicationScope{
+		applicationScope := aqua.ApplicationScope{
 			Name:               aquaScannerAccountName,
 			Description:        "Scanner scoped to " + namespacePrefix + "-* and DockerHub only.",
 			TechnicalLeadEmail: technicalLeadEmail,
 			NamespacePrefix:    namespacePrefix,
 		}
 
-		role := Role{
+		role := aqua.Role{
 			Name:             aquaScannerAccountName,
 			Description:      "AquaScannerAccount created Role to allow Scanning of resources scoped to " + namespacePrefix + "-* and DockerHub only.",
 			ApplicationScope: applicationScope,
 		}
 
-		applicationScopeErr := createAquaApplicationScope(ctrl.Log, applicationScope)
+		applicationScopeErr := aqua.CreateAquaApplicationScope(ctrl.Log, applicationScope)
 
 		if applicationScopeErr != nil {
 			ctrl.Log.Error(applicationScopeErr, "Failed to create application scope")
+			aquaScannerAccount.Status.CurrentState = "Failure"
+			aquaScannerAccount.Status.Timestamp = v1.Timestamp{Seconds: time.Now().Unix(), Nanos: int32(time.Now().UnixNano())}
+			updateErr = r.Status().Update(ctx, aquaScannerAccount)
+
+			if updateErr != nil {
+				ctrl.Log.Error(updateErr, "Failed to update aquaScannerAccount status")
+				return ctrl.Result{Requeue: true}, updateErr
+			}
 			return ctrl.Result{Requeue: true}, applicationScopeErr
 		}
 
-		roleErr := createAquaRole(ctrl.Log, role)
+		roleErr := aqua.CreateAquaRole(ctrl.Log, role)
 
 		if roleErr != nil {
 			ctrl.Log.Error(roleErr, "Failed to create role")
+			aquaScannerAccount.Status.CurrentState = "Failure"
+			aquaScannerAccount.Status.Timestamp = v1.Timestamp{Seconds: time.Now().Unix(), Nanos: int32(time.Now().UnixNano())}
+			updateErr = r.Status().Update(ctx, aquaScannerAccount)
+
+			if updateErr != nil {
+				ctrl.Log.Error(updateErr, "Failed to update aquaScannerAccount status")
+				return ctrl.Result{Requeue: true}, updateErr
+			}
+
 			return ctrl.Result{Requeue: true}, roleErr
 		}
 
-		config := generator.Config{
-			Length:                     16,
-			IncludeSymbols:             false,
-			IncludeNumbers:             true,
-			IncludeLowercaseLetters:    true,
-			IncludeUppercaseLetters:    true,
-			ExcludeSimilarCharacters:   true,
-			ExcludeAmbiguousCharacters: true,
+		var pwd *string
+
+		if aquaScannerAccount.Status.AccountSecret != "" {
+			pwd = &aquaScannerAccount.Status.AccountSecret
+		} else {
+			config := generator.Config{
+				Length:                     16,
+				IncludeSymbols:             false,
+				IncludeNumbers:             true,
+				IncludeLowercaseLetters:    true,
+				IncludeUppercaseLetters:    true,
+				ExcludeSimilarCharacters:   true,
+				ExcludeAmbiguousCharacters: true,
+			}
+			g, _ := generator.New(&config)
+
+			pwd, _ = g.Generate()
 		}
-		g, _ := generator.New(&config)
 
-		pwd, _ := g.Generate()
-
-		user := User{
+		user := aqua.User{
 			Name:     aquaScannerAccountName,
 			Password: *pwd,
 			Role:     role,
 		}
 
-		userErr := createAquaAccount(ctrl.Log, user)
+		aquaScannerAccount.Status.AccountName = user.Name
+		aquaScannerAccount.Status.AccountSecret = user.Password
+		aquaScannerAccount.Status.Timestamp = v1.Timestamp{Seconds: time.Now().Unix(), Nanos: int32(time.Now().UnixNano())}
+		updateErr = r.Status().Update(ctx, aquaScannerAccount)
+
+		if updateErr != nil {
+			ctrl.Log.Error(updateErr, "Failed to update aquaScannerAccount status")
+			return ctrl.Result{Requeue: true}, updateErr
+		}
+
+		userErr := aqua.CreateAquaAccount(ctrl.Log, user)
 
 		if userErr != nil {
 			ctrl.Log.Error(userErr, "Failed to create user")
+			aquaScannerAccount.Status.CurrentState = "Failure"
+			aquaScannerAccount.Status.Timestamp = v1.Timestamp{Seconds: time.Now().Unix(), Nanos: int32(time.Now().UnixNano())}
+			updateErr = r.Status().Update(ctx, aquaScannerAccount)
+
+			if updateErr != nil {
+				ctrl.Log.Error(updateErr, "Failed to update aquaScannerAccount status")
+				return ctrl.Result{Requeue: true}, updateErr
+			}
+
 			return ctrl.Result{Requeue: true}, userErr
 		}
 
 		// set status to Complete
-		aquaScannerAccount.Status.CurrentState = "Failure"
-		aquaScannerAccount.Status.AccountName = user.Name
-		aquaScannerAccount.Status.AccountSecret = user.Password
+		aquaScannerAccount.Status.CurrentState = "Complete"
 		aquaScannerAccount.Status.Timestamp = v1.Timestamp{Seconds: time.Now().Unix(), Nanos: int32(time.Now().UnixNano())}
 		updateErr = r.Status().Update(ctx, aquaScannerAccount)
 
@@ -268,20 +264,20 @@ func (r *AquaScannerAccountReconciler) SetupWithManager(mgr ctrl.Manager) error 
 
 func (r *AquaScannerAccountReconciler) finalizeAquaScannerAccount(reqLogger *log.DelegatingLogger, m *mamoadevopsgovbccav1alpha1.AquaScannerAccount, aquaScannerName string) error {
 
-	delAcctErr := deleteAquaAccount(ctrl.Log, aquaScannerName)
+	delAcctErr := aqua.DeleteAquaAccount(ctrl.Log, aquaScannerName)
 
 	if delAcctErr != nil {
 		return delAcctErr
 	}
 
-	delAppScopeErr := deleteAquaApplicationScope(ctrl.Log, aquaScannerName)
+	delAppScopeErr := aqua.DeleteAquaApplicationScope(ctrl.Log, aquaScannerName)
 
 	if delAppScopeErr != nil {
 		return delAppScopeErr
 	}
 	// delete application scope
 	// delete role
-	delRoleErr := deleteAquaRole(ctrl.Log, aquaScannerName)
+	delRoleErr := aqua.DeleteAquaRole(ctrl.Log, aquaScannerName)
 	if delRoleErr != nil {
 		return delRoleErr
 	}
@@ -289,307 +285,5 @@ func (r *AquaScannerAccountReconciler) finalizeAquaScannerAccount(reqLogger *log
 	// of finalizers include performing backups and deleting
 	// resources that are not owned by this CR, like a PVC.
 	reqLogger.Info("Successfully finalized AquaScannerAccount")
-	return nil
-}
-
-func doesAquaAccountAlreadyExist(reqLogger *log.DelegatingLogger, accountName string) (bool, error) {
-
-	aquaAuth := utils.GetAquaAuth()
-	jwt, jwtErr := aquaAuth.GetJWT()
-	if jwtErr != nil {
-		reqLogger.Error(jwtErr, "Failed to login to Aqua")
-		return false, jwtErr
-	}
-
-	reqLogger.Info("Checking if %v was created previously in aqua", accountName)
-	reqUrl := os.Getenv("AQUA_URL") + "/users/" + accountName
-	client := &http.Client{}
-	req, _ := http.NewRequest("GET", reqUrl, nil)
-	req.Header.Set("Authorization", "Bearer "+jwt)
-	req.Header.Set("Accept", "application/json")
-
-	res, err := client.Do(req)
-
-	if err != nil {
-		reqLogger.Error(err, "Failed request to GET %v from aqua", accountName)
-		return false, err
-	}
-
-	if res.StatusCode == 404 {
-		reqLogger.Info("User %v not found in aqua", accountName)
-		return false, nil
-	}
-
-	if res.StatusCode == 200 {
-		reqLogger.Info("User %v already exists in aqua", accountName)
-		return true, nil
-	}
-
-	errorMsg := "There was an issue making the request to GET user " + accountName + " from aqua"
-	qualifiedResource := schema.GroupResource{Group: mamoadevopsgovbccav1alpha1.GroupVersion.Group, Resource: "AquaScannerAccount"}
-
-	return false, errors.NewGenericServerResponse(res.StatusCode, "GET", qualifiedResource, "Generic Error", errorMsg, 10, true)
-}
-
-func deleteAquaAccount(reqLogger *log.DelegatingLogger, accountName string) error {
-	reqLogger.Info("Deleting user %v in aqua", accountName)
-
-	aquaAuth := utils.GetAquaAuth()
-	jwt, jwtErr := aquaAuth.GetJWT()
-	if jwtErr != nil {
-		reqLogger.Error(jwtErr, "Failed to login to Aqua")
-		return jwtErr
-	}
-
-	reqUrl := os.Getenv("AQUA_URL") + "/api/v1/users/" + accountName
-	client := &http.Client{}
-	req, _ := http.NewRequest("DELETE", reqUrl, nil)
-	req.Header.Set("Authorization", "Bearer "+jwt)
-	req.Header.Set("Accept", "application/json")
-
-	res, err := client.Do(req)
-
-	if err != nil {
-		reqLogger.Error(err, "Failed request to DELETE /api/v1/users %v from aqua", accountName)
-		return err
-	}
-
-	if res.StatusCode != 204 {
-		reqLogger.Error(err, "Failed to DELETE /api/v1/users %v from aqua", accountName)
-		return errors.NewBadRequest("Failed to DELETE user from aqua")
-	}
-	reqLogger.Info("User %v deleted", accountName)
-	return nil
-}
-
-func deleteAquaApplicationScope(reqLogger *log.DelegatingLogger, applicationScope string) error {
-	reqLogger.Info("Deleting applicationScope %v in aqua", applicationScope)
-
-	aquaAuth := utils.GetAquaAuth()
-	jwt, jwtErr := aquaAuth.GetJWT()
-	if jwtErr != nil {
-		reqLogger.Error(jwtErr, "Failed to login to Aqua")
-		return jwtErr
-	}
-
-	reqPayload, jsonErr := json.Marshal([]string{applicationScope})
-
-	if jsonErr != nil {
-		reqLogger.Error(jsonErr, "Failed to marshal json %v", []string{applicationScope})
-		return jsonErr
-	}
-
-	reqUrl := os.Getenv("AQUA_URL") + "/api/v2/access_management/scopes/delete"
-	client := &http.Client{}
-
-	req, _ := http.NewRequest("POST", reqUrl, bytes.NewBuffer(reqPayload))
-	req.Header.Set("Authorization", "Bearer "+jwt)
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Accept", "application/json")
-
-	res, err := client.Do(req)
-
-	if err != nil {
-		reqLogger.Error(err, "Failed request to POST to /api/v2/access_management/scopes/delete in aqua")
-		return err
-	}
-
-	if res.StatusCode != 204 {
-		e := errors.NewBadRequest(fmt.Sprintf("Error: Could not delete application scope, the response status from aqua was %v", res.StatusCode))
-		return e
-	}
-	return nil
-}
-
-func deleteAquaRole(reqLogger *log.DelegatingLogger, role string) error {
-	reqLogger.Info("Deleting role %v in aqua", role)
-
-	aquaAuth := utils.GetAquaAuth()
-	jwt, jwtErr := aquaAuth.GetJWT()
-	if jwtErr != nil {
-		reqLogger.Error(jwtErr, "Failed to login to Aqua")
-		return jwtErr
-	}
-
-	reqUrl := os.Getenv("AQUA_URL") + "/api/v2/access_management/roles/" + role
-	client := &http.Client{}
-
-	req, _ := http.NewRequest("DELETE", reqUrl, nil)
-	req.Header.Set("Authorization", "Bearer "+jwt)
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Accept", "application/json")
-
-	res, err := client.Do(req)
-
-	if err != nil {
-		reqLogger.Error(err, "Failed request to DELETE to /api/v2/access_management/roles/ in aqua")
-		return err
-	}
-
-	if res.StatusCode != 204 {
-		e := errors.NewBadRequest(fmt.Sprintf("Error: Could not delete role, the response status from aqua was %v", res.StatusCode))
-		return e
-	}
-	return nil
-}
-
-func createAquaRole(reqLogger *log.DelegatingLogger, role Role) error {
-	reqLogger.Info("Creating Role %v in aqua", role.Name)
-
-	aquaAuth := utils.GetAquaAuth()
-	jwt, jwtErr := aquaAuth.GetJWT()
-	if jwtErr != nil {
-		reqLogger.Error(jwtErr, "Failed to login to Aqua")
-		return jwtErr
-	}
-
-	b, fileErr := ioutil.ReadFile("../templates/Role.json.tmpl")
-
-	if fileErr != nil {
-		reqLogger.Error(fileErr, "Failed to read template file Role.json.tmpl")
-		return fileErr
-	}
-
-	ut, templateErr := template.New("Role").Parse(string(b))
-
-	if templateErr != nil {
-		reqLogger.Error(templateErr, "Failed to parse template file Role.json.tmpl")
-		return templateErr
-	}
-
-	var roleBuffer bytes.Buffer
-	ut.Execute(&roleBuffer, role)
-
-	reqUrl := os.Getenv("AQUA_URL") + " /api/v2/access_management/roles"
-	client := &http.Client{}
-	req, _ := http.NewRequest("POST", reqUrl, &roleBuffer)
-	req.Header.Set("Authorization", "Bearer "+jwt)
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Accept", "application/json")
-
-	res, err := client.Do(req)
-
-	if err != nil {
-		reqLogger.Error(err, "Failed request to POST to  /api/v2/access_management/roles in aqua")
-		return err
-	}
-	defer res.Body.Close()
-
-	var jsonData aquaResponseJson
-	body, _ := ioutil.ReadAll(res.Body)
-
-	json.Unmarshal(body, &jsonData)
-
-	if res.StatusCode == 404 && strings.Contains(jsonData.Message, "role "+role.Name+" already exists") || res.StatusCode == 201 {
-		return nil
-	} else {
-		e := errors.NewBadRequest(fmt.Sprintf("Error: Could not create role, the response status from aqua was %v", res.StatusCode))
-
-		reqLogger.Error(e, "Unable to create Role")
-		return e
-	}
-}
-
-func createAquaApplicationScope(reqLogger *log.DelegatingLogger, appScope ApplicationScope) error {
-	reqLogger.Info("Creating applicationScope %v-* in aqua", appScope.NamespacePrefix)
-	aquaAuth := utils.GetAquaAuth()
-	jwt, jwtErr := aquaAuth.GetJWT()
-	if jwtErr != nil {
-		reqLogger.Error(jwtErr, "Failed to login to Aqua")
-		return jwtErr
-	}
-
-	b, fileErr := ioutil.ReadFile("../templates/ApplicationScope.json.tmpl")
-
-	if fileErr != nil {
-		reqLogger.Error(fileErr, "Failed to read template file ApplicationScope.json.tmpl")
-		return fileErr
-	}
-
-	ut, templateErr := template.New("ApplicationScope").Parse(string(b))
-
-	if templateErr != nil {
-		reqLogger.Error(templateErr, "Failed to parse template file ApplicationScope.json.tmpl")
-		return templateErr
-	}
-
-	var appScopeBuffer bytes.Buffer
-	ut.Execute(&appScopeBuffer, appScope)
-
-	reqUrl := os.Getenv("AQUA_URL") + " /api/v2/access_management/scopes"
-	client := &http.Client{}
-	req, _ := http.NewRequest("POST", reqUrl, &appScopeBuffer)
-	req.Header.Set("Authorization", "Bearer "+jwt)
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Accept", "application/json")
-
-	res, err := client.Do(req)
-
-	if err != nil {
-		reqLogger.Error(err, "Failed request to POST to  /api/v2/access_management/scopes in aqua")
-		return err
-	}
-	defer res.Body.Close()
-
-	var jsonData aquaResponseJson
-	body, _ := ioutil.ReadAll(res.Body)
-
-	json.Unmarshal(body, &jsonData)
-
-	if res.StatusCode == 404 && strings.Contains(jsonData.Message, "application scope "+appScope.Name+" already exists") || res.StatusCode == 201 {
-		return nil
-	} else {
-		e := errors.NewBadRequest(fmt.Sprintf("Error: Could not create ApplicationScope, the response status from aqua was %v", res.StatusCode))
-
-		reqLogger.Error(e, "Unable to create ApplicationScope")
-		return e
-	}
-}
-
-func createAquaAccount(reqLogger *log.DelegatingLogger, user User) error {
-	reqLogger.Info("Creating user %v in aqua", user.Name)
-
-	aquaAuth := utils.GetAquaAuth()
-	jwt, jwtErr := aquaAuth.GetJWT()
-	if jwtErr != nil {
-		reqLogger.Error(jwtErr, "Failed to login to Aqua")
-		return jwtErr
-	}
-
-	b, fileErr := ioutil.ReadFile("../templates/User.json.tmpl")
-
-	if fileErr != nil {
-		reqLogger.Error(fileErr, "Failed to read template file User.json.tmpl")
-		return fileErr
-	}
-
-	ut, templateErr := template.New("User").Parse(string(b))
-
-	if templateErr != nil {
-		reqLogger.Error(templateErr, "Failed to parse template file User.json.tmpl")
-		return templateErr
-	}
-
-	var userBuffer bytes.Buffer
-	ut.Execute(&userBuffer, user)
-
-	reqUrl := os.Getenv("AQUA_URL") + "/api/v1/users/"
-	client := &http.Client{}
-	req, _ := http.NewRequest("POST", reqUrl, &userBuffer)
-	req.Header.Set("Authorization", "Bearer "+jwt)
-	req.Header.Set("Content-Type", "application/json")
-
-	res, err := client.Do(req)
-
-	if err != nil {
-		reqLogger.Error(err, "Failed request to POST /api/v1/ %v from aqua", user.Name)
-		return err
-	}
-
-	if res.StatusCode != 204 {
-		reqLogger.Error(err, "Failed to POST %v from aqua. Status code is %v", user.Name, res.StatusCode)
-		return errors.NewBadRequest("Failed to POST user from aqua")
-	}
-	reqLogger.Info("User %v created in aqua", user.Name)
 	return nil
 }
